Somechanges

+++ api/controllers/users.js
@@ -1,7 +1,8 @@
-const { User, AccessToken } = require('../../models');
+const { User, Profile } = require('../../models');
 const bcrypt = require('bcrypt');
 const generateNewToken = require('../utils/generateAccessToken');
 const storeAccessToken = require('../utils/storeAccessToken');
+const createProfile = require('../utils/createProfile');
 
 exports.createUser = async (req, res, next) => {
   const { email, password } = req.body;
@@ -17,16 +18,18 @@ exports.createUser = async (req, res, next) => {
       res.status(500).json({ message: err.message });
     } else {
       try {
-        await User.create({
+        const user = await User.create({
           email: email,
           password: hash,
+          role: 'customer',
         });
+        await createProfile(user.id);
+        res.status(200).json({ message: 'user created successfully' });
       } catch (err) {
-        res.status(500).json({ message: err.message });
+        return res.status(500).json({ message: err.message });
       }
     }
   });
-  res.status(200).json({ message: 'user created successfully' });
 };
 
 exports.signin = async (req, res, next) => {
@@ -69,18 +72,43 @@ exports.signin = async (req, res, next) => {
   }
 };
 
-exports.deleteUser = async (req, res, next) => {
-  const userId = req.params.id;
-  const userCheck = await User.findOne({ where: { userId } });
-  if (!userCheck) {
-    return res.status(400).json({
-      message: 'user does not exist',
-    });
+exports.updateProfile = async (req, res, next) => {
+  const userId = req.userData.id;
+  console.log(req.userData);
+  const profile = await Profile.findOne({ where: { userId } });
+  const { firstName, lastName, bio } = req.body;
+  if (profile) {
+    try {
+      profile.firstName = firstName;
+      profile.lastName = lastName;
+      profile.bio = bio;
+      profile.save();
+      res.status(200).json(profile);
+    } catch (err) {
+      res.status(500).json({ message: err.message });
+    }
+  } else {
+    res.status(404).json({ message: 'profile not found' });
   }
-  try {
-    await User.destroy({ where: { userId } });
-    res.status(200).json({ message: 'user deleted successfully' });
-  } catch (err) {
-    res.status(500).json({ message: err.message });
+};
+
+exports.deleteUser = async (req, res, next) => {
+  const userRole = req.userData.role;
+  if (userRole === 'admin' || userRole === 'supervisor') {
+    const userId = req.params.id;
+    const userCheck = await User.findOne({ where: { userId } });
+    if (!userCheck) {
+      return res.status(400).json({
+        message: 'user does not exist',
+      });
+    }
+    try {
+      await User.destroy({ where: { userId } });
+      res.status(200).json({ message: 'user deleted successfully' });
+    } catch (err) {
+      res.status(500).json({ message: err.message });
+    }
+  } else {
+    res.status(401).json({ message: "you're  unauthorized to do this action" });
   }
 };
